refactor(pipeline-cli): single-source the gh-IO seam; migrate verdict onto it (#3572)
Extract the runGh / resolveRepo / authorizedAuthors / RawComment idiom the Tracker
service, epic-lock, and verdict each re-copied into one shared `tracker/gh-io.ts`
seam (the #3262 AC 5 same-commit-adoption follow-up, ADR 0190). The seam is the
gh-over-REST plumbing only — spawn, JSON parse, ADR 0062 repo resolution, the ADR
0055 write+ ACL gate, the raw comment shape, and the generic comment arg builders
(list/post/patch/delete/get-body + whoami); each consumer keeps its own domain core.

tracker.ts and verdict/github.ts now import the seam and delete their private copies.
The typed failures (GhCommandError / GhParseError / RepoResolutionError) collapse to
one `@kampus/gh-io/*` namespace, re-exported from each module so callers/tests keep
importing them unchanged. No observable behavior change — all existing tests pass.
