+++ packages/pipeline-cli/src/tools/tracker/tracker.ts
@@ -33,9 +33,9 @@
  * IO-free, ADR-0040-tested core (`../epic-lock/claim-resolution.ts`) rather than
  * re-derived — the single source of "who owns this claim" (gh-issue-intake-formats.md §7).
  */
-import {Context, Effect, Layer, Stream} from "effect";
+import {Context, Effect, Layer} from "effect";
 import * as Schema from "effect/Schema";
-import {ChildProcess, ChildProcessSpawner} from "effect/unstable/process";
+import {ChildProcessSpawner} from "effect/unstable/process";
 import {type ClaimComment, type ClaimWinner, resolveWinner} from "../epic-lock/claim-resolution.ts";
 // The ADR-0058 verdict-marker grammar + emission guard is the SINGLE SOURCE — reused, never
 // re-derived, exactly as the claim decision reuses `epic-lock/claim-resolution.ts`. `postVerdict`
@@ -48,33 +48,31 @@ import {
 	namespaceRe,
 	type VerdictGate,
 } from "../verdict/verdict-match.ts";
+// The `gh api` REST IO seam (runGh / resolveRepo / authorizedAuthors / RawComment) is the SINGLE
+// SOURCE this service, `epic-lock/github.ts`, and `verdict/github.ts` share — no longer re-copied
+// per consumer (#3262 AC 5). The error types are re-exported below so callers/tests keep importing
+// them from this module.
+import {
+	authorizedAuthors,
+	decodeComments,
+	deleteCommentArgs,
+	type GhCommandError,
+	GhParseError,
+	getCommentBodyArgs,
+	json,
+	listCommentsArgs,
+	patchCommentArgs,
+	postCommentArgs,
+	type RawComment,
+	type RepoResolutionError,
+	resolveRepo,
+	runGh,
+	whoAmIArgs,
+} from "./gh-io.ts";
 
-/** A `gh` invocation exited non-zero (auth, not-found, rate-limit, …). */
-export class GhCommandError extends Schema.TaggedErrorClass<GhCommandError>()(
-	"@kampus/tracker/GhCommandError",
-	{
-		args: Schema.Array(Schema.String),
-		exitCode: Schema.Number,
-		stderr: Schema.String,
-	},
-) {}
-
-/** `gh` output was not the JSON the loader expected. */
-export class GhParseError extends Schema.TaggedErrorClass<GhParseError>()(
-	"@kampus/tracker/GhParseError",
-	{
-		args: Schema.Array(Schema.String),
-		message: Schema.String,
-	},
-) {}
-
-/** No `owner/name` target repo could be resolved (no env override, no current repo). */
-export class RepoResolutionError extends Schema.TaggedErrorClass<RepoResolutionError>()(
-	"@kampus/tracker/RepoResolutionError",
-	{
-		message: Schema.String,
-	},
-) {}
+// Re-export the shared IO seam's typed failures so callers/tests keep importing them from this
+// service module — the single-source classes now live in `gh-io.ts` (#3262 AC 5).
+export {GhCommandError, GhParseError, RepoResolutionError} from "./gh-io.ts";
 
 /**
  * A malformed verb judgment the caller must fix — e.g. `postVerdict` handed an unknown gate, or a
@@ -250,136 +248,9 @@ export type GraduateResult = {
 	readonly state: string;
 };
 
-const collect = (stream: Stream.Stream<Uint8Array, unknown>): Effect.Effect<string> =>
-	Stream.decodeText(stream).pipe(
-		Stream.mkString,
-		Effect.orElseSucceed(() => ""),
-	);
-
-/**
- * Run `gh <args>` and return stdout, failing `GhCommandError` on a non-zero exit — the
- * same direct-spawn shape `epic-lock`/`verdict` use so an exit code + stderr lower into a
- * typed error rather than a throw. A spawn/IO `PlatformError` (e.g. `gh` off PATH) folds
- * into the same typed error as exit `-1`.
- */
-const runGh = Effect.fn("Tracker.runGh")(
-	function* (args: ReadonlyArray<string>) {
-		const handle = yield* ChildProcess.make("gh", args);
-		const [stdout, stderr, exitCode] = yield* Effect.all(
-			[collect(handle.stdout), collect(handle.stderr), handle.exitCode],
-			{concurrency: "unbounded"},
-		);
-		if (exitCode !== 0) {
-			return yield* new GhCommandError({args, exitCode, stderr});
-		}
-		return stdout;
-	},
-	Effect.scoped,
-	(effect, args) =>
-		Effect.catchTag(
-			effect,
-			"PlatformError",
-			(cause) => new GhCommandError({args, exitCode: -1, stderr: cause.message}),
-		),
-);
-
-const parseJson = (
-	args: ReadonlyArray<string>,
-	raw: string,
-): Effect.Effect<unknown, GhParseError> =>
-	Effect.try({
-		try: () => JSON.parse(raw) as unknown,
-		catch: (cause) =>
-			new GhParseError({args, message: cause instanceof Error ? cause.message : String(cause)}),
-	});
-
-const json = Effect.fn("Tracker.json")(function* (args: ReadonlyArray<string>) {
-	return yield* parseJson(args, yield* runGh(args));
-});
-
-const REPO_RE = /^[^/\s]+\/[^/\s]+$/;
-
-/**
- * Resolve the target repo (`owner/name`) once, per ADR 0062 §1, in order:
- * `CLAUDE_PIPELINE_REPO` → `GITHUB_REPOSITORY` (CI) → `gh repo view`. Never silently
- * defaults: with no env and no resolvable current repo it fails `RepoResolutionError`, so
- * a foreign install can't accidentally operate on phoenix.
- */
-const resolveRepo = Effect.fn("Tracker.resolveRepo")(function* () {
-	const fromEnv = process.env.CLAUDE_PIPELINE_REPO ?? process.env.GITHUB_REPOSITORY;
-	if (fromEnv && REPO_RE.test(fromEnv.trim())) {
-		return fromEnv.trim();
-	}
-	const viewed = yield* runGh([
-		"repo",
-		"view",
-		"--json",
-		"nameWithOwner",
-		"-q",
-		".nameWithOwner",
-	]).pipe(
-		Effect.map((out) => out.trim()),
-		Effect.catchTag("@kampus/tracker/GhCommandError", () => Effect.succeed("")),
-	);
-	if (REPO_RE.test(viewed)) {
-		return viewed;
-	}
-	return yield* new RepoResolutionError({
-		message:
-			"could not resolve a target repo: set CLAUDE_PIPELINE_REPO (or GITHUB_REPOSITORY), " +
-			"or run inside a git repo whose origin `gh repo view` can read",
-	});
-});
-
-// REST-only arg builders — never GraphQL (broken on the kamp-us org).
-
-const listCommentsArgs = (repo: string, target: TargetId): ReadonlyArray<string> => [
-	"api",
-	"--paginate",
-	`repos/${repo}/issues/${target}/comments?per_page=100`,
-];
-
-// A generic comment POST — the claim marker and the verdict marker are both a body POST that
-// returns the new comment id; single-sourced so the two consumers can't drift.
-const postCommentArgs = (repo: string, target: TargetId, body: string): ReadonlyArray<string> => [
-	"api",
-	"-X",
-	"POST",
-	`repos/${repo}/issues/${target}/comments`,
-	"-f",
-	`body=${body}`,
-	"--jq",
-	".id",
-];
-
-const patchCommentArgs = (repo: string, id: number, body: string): ReadonlyArray<string> => [
-	"api",
-	"-X",
-	"PATCH",
-	`repos/${repo}/issues/comments/${id}`,
-	"-f",
-	`body=${body}`,
-	"--jq",
-	".id",
-];
-
-// The read-back GET for `postVerdict`'s self-verify (#3019): re-fetch the single comment we just
-// upserted and return its LANDED body, so the marker/leak-clean shape is re-checked as it landed.
-const getCommentBodyArgs = (repo: string, id: number): ReadonlyArray<string> => [
-	"api",
-	`repos/${repo}/issues/comments/${id}`,
-	"--jq",
-	".body",
-];
-
-const whoAmIArgs: ReadonlyArray<string> = ["api", "user", "--jq", ".login"];
-
-const deleteCommentArgs = (repo: string, id: number): ReadonlyArray<string> => [
-	"api",
-	"-X",
-	"DELETE",
-	`repos/${repo}/issues/comments/${id}`,
-];
+// Tracker-specific REST arg builders — never GraphQL (broken on the kamp-us org). The generic
+// comment IO (list/post/patch/delete/get-body) and the `whoami` probe are the shared `gh-io.ts`
+// seam; only the create/label/close envelopes unique to the tracker's verbs live here.
 
 // Create a new issue. `title` / `body` / each `label` pass as by-value POST fields through
 // the direct `gh` spawn — no shell, no heredoc, no `-f body=@file`, so the whole local-path
@@ -425,13 +296,6 @@ const closeCompletedArgs = (repo: string, target: TargetId): ReadonlyArray<strin
 	"state_reason=completed",
 ];
 
-const permissionArgs = (repo: string, login: string): ReadonlyArray<string> => [
-	"api",
-	`repos/${repo}/collaborators/${login}/permission`,
-	"--jq",
-	".permission",
-];
-
 const addLabelsArgs = (
 	repo: string,
 	target: TargetId,
@@ -458,15 +322,6 @@ const listLabelsArgs = (repo: string, target: TargetId): ReadonlyArray<string> =
 	"--paginate",
 	`repos/${repo}/issues/${target}/labels?per_page=100`,
 ];
-
-/** A raw comment as the issues/comments endpoint returns it; only these fields are read. */
-const RawComment = Schema.Struct({
-	id: Schema.Number,
-	created_at: Schema.String,
-	body: Schema.optionalKey(Schema.NullOr(Schema.String)),
-	user: Schema.NullOr(Schema.Struct({login: Schema.String})),
-});
-const decodeComments = Schema.decodeUnknownEffect(Schema.Array(RawComment));
 
 /** A raw label as the issues/labels endpoint returns it; only the name is read. */
 const RawLabel = Schema.Struct({name: Schema.String});
@@ -509,35 +364,6 @@ const listClaimComments = Effect.fn("Tracker.listClaimComments")(function* (
 	const args = listCommentsArgs(repo, target);
 	const raw = yield* decodeComments(yield* json(args));
 	return raw.map(toClaimComment);
-});
-
-/**
- * The write+ collaborator subset of `logins` — the ADR 0055 trust root, applied *inside*
- * the live impl so it never leaks into the `Tracker` interface (local-markdown has no ACL).
- * A non-`admin|maintain|write` permission, or any `gh` fault on the probe (a
- * non-collaborator commonly 404s), drops the login — a forged claim never enters the set
- * the pure core resolves over.
- */
-const authorizedAuthors = Effect.fn("Tracker.authorizedAuthors")(function* (
-	repo: string,
-	logins: ReadonlyArray<string>,
-) {
-	const results = yield* Effect.forEach(
-		logins,
-		(login) =>
-			runGh(permissionArgs(repo, login)).pipe(
-				Effect.map((out) => ({login, permission: out.trim()})),
-				Effect.catchTag("@kampus/tracker/GhCommandError", () =>
-					Effect.succeed({login, permission: "none"}),
-				),
-			),
-		{concurrency: "unbounded"},
-	);
-	return results
-		.filter(
-			(r) => r.permission === "admin" || r.permission === "maintain" || r.permission === "write",
-		)
-		.map((r) => r.login);
 });
 
 const resolveAuthorizedWinner = Effect.fn("Tracker.resolveAuthorizedWinner")(function* (
@@ -608,7 +434,7 @@ const applyTriage = Effect.fn("Tracker.applyTriage")(function* (
 	const add = [`type:${judgment.type}`, judgment.priority, `${LABEL_STATUS_PREFIX}${status}`];
 	yield* runGh(addLabelsArgs(repo, target, add));
 	yield* runGh(removeLabelArgs(repo, target, `${LABEL_STATUS_PREFIX}${QUEUE_STATUS}`)).pipe(
-		Effect.catchTag("@kampus/tracker/GhCommandError", () => Effect.succeed("")),
+		Effect.catchTag("@kampus/gh-io/GhCommandError", () => Effect.succeed("")),
 	);
 	const labels = yield* decodeLabels(yield* json(listLabelsArgs(repo, target)));
 	const landedStatus = labels
@@ -657,7 +483,7 @@ const verifyLanded = Effect.fn("Tracker.verifyLanded")(function* (
 ) {
 	const landed = yield* runGh(getCommentBodyArgs(repo, id)).pipe(
 		Effect.catchTag(
-			"@kampus/tracker/GhCommandError",
+			"@kampus/gh-io/GhCommandError",
 			(cause) =>
 				new TrackerVerifyError({
 					message: `could not re-fetch the just-posted verdict comment #${id} to self-verify it (${cause.stderr.trim() || `exit ${cause.exitCode}`}) — refusing to report success on an unverifiable post (#3019)`,
